feat: added support for image signatures
This should be helpful for using seals, stamps and chops on documents.
The implementation is pretty straightforward, uploaded images are rendered
directly onto the canvas.

It added a button to trigger the upload and has support for dragging and dropping.

Split the field size into default and min variables so we can allow it to default
to a bigger size while still allowing the small minimum size.